Fixed xxxMemset tests

+++ tests/simple/gmac/cuda/gmacMemset.cu
@@ -4,16 +4,16 @@
 const size_t size = 4 * 1024 * 1024;
 const size_t blockSize = 512;
 
-__global__ void reset(long *a, long v)
+__global__ void reset(unsigned *a, unsigned v)
 {
 	int i = threadIdx.x + blockIdx.x * blockDim.x;
 	if(i >= size) return;
 	a[i] += v;
 }
 
-int check(long *ptr, int s)
+int check(unsigned *ptr, unsigned s)
 {
-	int a = 0;
+	unsigned a = 0;
 	for(unsigned i = 0; i < size; i++)
 		a += ptr[i];
 	return a - s;
@@ -21,8 +21,9 @@ int check(long *ptr, int s)
 
 int main(int argc, char *argv[])
 {
-	long *ptr;
-	assert(gmacMalloc((void **)&ptr, size * sizeof(long)) == gmacSuccess);
+	unsigned *ptr;
+
+	assert(gmacMalloc((void **)&ptr, size * sizeof(unsigned)) == gmacSuccess);
 
 	// Call the kernel
 	dim3 Db(blockSize);
@@ -30,24 +31,24 @@ int main(int argc, char *argv[])
 	if(size % blockSize) Db.x++;
 
 	fprintf(stderr,"Test GMAC full memset: ");
-    gmacMemset(ptr, 0, size * sizeof(long));
+    gmacMemset(ptr, 0, size * sizeof(unsigned));
 
 	reset<<<Dg, Db>>>(gmacPtr(ptr), 1);
     gmacThreadSynchronize();
 	fprintf(stderr,"%d\n", check(ptr, size));
 
 	fprintf(stderr, "Test GMAC partial memset: ");
-	gmacMemset(&ptr[size / 8], 0, 3 * size / 4 * sizeof(long));
+	gmacMemset(&ptr[size / 8], 0, 3 * size / 4 * sizeof(unsigned));
 	fprintf(stderr,"%d\n", check(ptr, size / 4));
 
 	fprintf(stderr,"Test STDC full memset: ");
-    memset(ptr, 0, size * sizeof(long));
+    memset(ptr, 0, size * sizeof(unsigned));
 	reset<<<Dg, Db>>>(gmacPtr(ptr), 1);
     gmacThreadSynchronize();
 	fprintf(stderr,"%d\n", check(ptr, size));
 
 	fprintf(stderr, "Test STDC partial memset: ");
-	memset(&ptr[size / 8], 0, 3 * size / 4 * sizeof(long));
+	memset(&ptr[size / 8], 0, 3 * size / 4 * sizeof(unsigned));
 	fprintf(stderr,"%d\n", check(ptr, size / 4));
 
 	gmacFree(ptr);
